Section headers now end code cells (#8127)
- Addresses https://github.com/posit-dev/positron/issues/7974
- Closes #8021

This is an internal PR applying the commits from #8021, because of the
problems outlined in #6628.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Section headers in `.R` files (from special comments like `# A section
----`) now act as end boundaries for code cells, thanks to @kylebutts

#### Bug Fixes

- N/A


### QA Notes

If we have an `.R` file such as this:

```r
# A section but not a code cell ----
1

## A subsection but not a code cell ----
2

# %% A code cell
3

# A new section but not a code cell ----
4

#+ A new code cell
5

# Another new section but not a code cell ----
6
```

we should have code cells for 3 and 5, but not for 1, 2, 4, or 6.

---------

Signed-off-by: Wasim Lorgat <mwlorgat@gmail.com>
Co-authored-by: Kyle Butts <buttskyle96@gmail.com>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>