Respect selected provider when using Apply in Editor (#8158)
This addresses an issue in which _Apply in Editor_ does not respect the
current chat model and just picks the first registered model to generate
the mapped edits. This can be a particular problem with the test
echo/error models since they do not generate responses we can parse as
edits.

The fix is to track the model ID used for the edit session in the chat
participant. This enables us to use the model of the chat session from
which the code block originated to generate the edits.

It also adds more safeguards to this scenario to make the error less
inscrutable:

- the model is now instructed even more clearly about what to return
(suggestion from @georgestagg)
- if the model returns just one edit and forgets to wrap it in an array,
we wrap it on the model's behalf
- if the model returns nonsense, it is logged with a warning instead of
generating an exception

Addresses https://github.com/posit-dev/positron/issues/7188

### QA Notes

To test this, register the Echo or Error provider _first_, then
Anthropic.

Note that because this change works by tracking the active edit session,
it only works if you _Apply in Editor_ in the same Positron session in
which the code was generated. That is, if you reload/restart and try to
_Apply in Editor_ a code chunk generated from before the reload/restart,
it's a new session, and _Apply in Editor_ will have to guess the model
as before.

---------

Signed-off-by: Jonathan <jonathan@posit.co>
Co-authored-by: Wasim Lorgat <mwlorgat@gmail.com>