Updated screenshot (#8286)
Due to https://github.com/posit-dev/positron/pull/8258 , the screenshots for the plots tests needed to be updated.

Since MacOS is only run nightly on release builds, we had to wait for it to get this last screenshot.
 
### QA Notes
We can't test locally as the screenshots are often different in CI.

<!--
  Positron team members: please add relevant e2e test tags, so the tests can be
  run when you open this pull request.

  - Instructions: https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
  - Available tags: https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->