Support functions in `View()` (#8262)
Addresses posit-dev/positron#2945.
Companion PR for https://github.com/posit-dev/ark/pull/848.

The main change needed on the frontend side is to detect non-file
schemes in the `OpenEditor` event handler of the UI comm. This way we
preserve `ark:` URIs, which allows Ark to request the frontend to open
documents via its own vdoc provider.

I also improved vdoc provider errors to be more user friendly. The error
is mentioned in the virtual editor and the error message is logged.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- `View()` now supports function objects (posit-dev/positron#2945). It
takes you directly to the original source, if the function has source
references, or to a virtual document containing the deparsed function
code otherwise.

- Functions are now also viewable from the Variables pane. Click on the
"view" button on the right hand side of a variable entry to view its
source definition in an editor.


#### Bug Fixes

- N/A


### QA Notes

See companion PR: https://github.com/posit-dev/ark/pull/848.

<!--
Positron team members: please add relevant e2e test tags, so the tests
can be
  run when you open this pull request.

- Instructions:
https://github.com/posit-dev/positron/blob/main/test/e2e/README.md#pull-requests-and-test-tags
- Available tags:
https://github.com/posit-dev/positron/blob/main/test/e2e/infra/test-runner/test-tags.ts
-->


<!--
  Add additional information for QA on how to validate the change,
  paying special attention to the level of risk, adjacent areas that
  could be affected by the change, and any important contextual
  information not present in the linked issues.
-->

@:variables