python: unskip plotnine test (#8288)
Closes #8239. Breaking change in `plotnine` dependency -> `statsmodels`
dependency -> `scipy` 🐛 . Pinning to earlier versions until
`statsmodels` does a new release.

⏫ specifically choosing to use closing keywords since green CI is all we
need to verify we are back! Added a note in
https://github.com/posit-dev/positron/issues/8184 to check if we can
unpin this.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #8239 Re-enable `plotnine` test in Python CI.


### QA Notes

Green CI