Persist session renames after reload for web builds (#8304)
Addresses #7692 

This PR addresses the session rename bug found on web builds of
Positron.

If a user renamed a session and then ran the command `Developer: Reload
Window` or clicked the "Reload page" button for the browser tab, the
session name would display the default name the session started with
(the runtime name) instead of the name provided by the user. This bug is
not present in desktop builds.

The reason for this difference in behavior is due to how web builds
handle shutdown events (`BrowserLifecycleService` vs the
`NativeLifecycleService`). (1) The `BrowserLifecycleService` always
sends a shutdown reason of `QUIT` when firing the `onBeforeShutdown`
event unlike the `NativeLifecycleService`. In the
`runtimeStartupService` we listen to `onBeforeShutdown` and save session
data only if the shutdown reason is set to `RELOAD` (which is never the
case for the web build). We do this to avoid the UX experience described
next. (2) The browser lifecycle doesn't support async operations during
shutdown. The service is hardcoded to veto all async shutdown operations
and show native dialog with a warning. The implementation details can be
found in `BrowserLifecycleService.onBeforeUnload` and
`BrowserLifecycleService.doShutdown`

The `runtimeStart` service now listens for session name changes and
saves the workspace sessions when that event is triggered.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Persist session renames in UI after reloads for web builds (#7692)

### QA Notes

Tested on workbench with local positron build per the instructions here:
https://github.com/posit-dev/positron-builds/tree/main/dev/workbench

@:sessions @:web

### Screenshots



https://github.com/user-attachments/assets/46defe41-04fd-4d5f-9703-038040452e69