e2e test: refactor data explorer pandas test for independence (#8340)
### Summary
Recently I updated the `data-explorer-python-polars.test.ts` to have
independent tests - this is a continuation of clean up doing the same
for `data-explorer-python-pandas.test.ts`. Each test in this spec is
easier to read/maintain and can be run independently w/o dependencies.

### QA Notes

@:data-explorer @:web @:win

Also kicked off the full suite here:
https://github.com/posit-dev/positron/actions/runs/15935116650