Improve keyboard navigation for Plots (#8369)
Quick change to support arrows to navigate the Plots history. 

Addresses https://github.com/posit-dev/positron/issues/8327.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Improved keyboard navigation in the Plots pane; arrow keys now
navigate the plot history. (#8327)

#### Bug Fixes

- N/A


### QA Notes

The arrow keys aren't scoped to the orientation of the filmstrip -- that
is, "up" and "left" will take you to the previous plot, and "down" or
"right" to the next one, regardless of whether the filmstrip is on the
top or bottom.

tags: `@:plots`