Offer an affordance for interrupting the session in order to reconnect (#8384)
This change makes it possible to interrupt a long-running computation in
order to reconnect to an R or Python session. In most cases, these
computations will eventually finish, but some things that make the
session busy (e.g. Shiny apps) literally run until they are interrupted.

The fix is to add a Cancel button to the reconnect notification:

<img width="467" alt="image"
src="https://github.com/user-attachments/assets/b758077c-7b9c-49b0-bfc2-8d882e66f4a2"
/>

When invoked, this Cancel button offers a way to interrupt the session:

<img width="413" alt="image"
src="https://github.com/user-attachments/assets/44b0f6c0-d8c8-4e55-838a-f45c81c06c95"
/>

Addresses https://github.com/posit-dev/positron/issues/8041.

In an ideal world, we would be able to reconnect even if the session is
busy. However, a lot of the operations we perform in order to connect
the sessions to Positron require that we send messages to the shell
socket, so they won't work while the kernel is busy. See #7447 for some
infrastructure that would help us mitigate this.

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- It is now possible to interrupt R and Python sessions in order to
reconnect to them (#8041)

#### Bug Fixes

- N/A


### QA Notes

You can use `Sys.sleep(500)` or something to simulate a long-running
operation in addition to the Shiny app example in the issues