Don't actually capture any of the line we use as a signal to stop the `@examples` section with
It seems like if we do this, then the highlighter will just start from the top again and apply the highlight rules like normal. We use lookaheads to avoid actually capturing any of the text we are using as our signal. The only thing we end up capturing is possibly some leading whitespace, which I think is okay.