Add LLM tools for R package help & vignettes (#8421)
This change adds 4 new LLM tools that can be used to read an R package's
help and vignettes.

<img width="415" alt="image"
src="https://github.com/user-attachments/assets/bd404fde-e160-4e19-ac4f-c9b944ee6ac3"
/>

Requires https://github.com/posit-dev/ark/pull/868 on the Ark side.

Addresses https://github.com/posit-dev/positron/issues/8016 (for R)


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- Positron Assistant can now reference R help pages and package
vignettes (#8016)

#### Bug Fixes

- N/A


### QA Notes

The LLM has been instructed to use help when it needs to, but a fun
thing about LLMs is that they mostly don't know what they don't know.
Asking for help with a package that is newer than the LLM's knowledge
cutoff date is a good way to get it to fall back to tool calls to look
up help; if you ask it about common usage for long-standing base R
functions, it is not likely to need help pages. You can also just tell
it to look up a help page or vignette and it'll do the right thing.