tell users to install numpy for polars histograms (#8444)
supersedes #8392 

This is motivated by the fact all the data explorer Python code assumes
you have numpy installed, but numpy is not actually required for polars.
Since the histogram method we would need to use for polars is
[unstable](https://docs.pola.rs/api/python/stable/reference/series/api/polars.Series.hist.html),
and we do not want to be the keepers/implementers of this type of code,
we'll just tell users to install numpy for now if they want histograms
in the data explorer.

Future work to move to polars-native histograms tracked here:
https://github.com/posit-dev/positron/issues/8446

### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- #8203 Inform users to install `numpy` if they want histograms in the
Data Explorer for `polars` data frames.


### QA Notes


```python
# do NOT have numpy installed

import polars as pl

sample_data_error = {"float": 3.14, "int": 1}
df = pl.DataFrame(sample_data_error)

%view df
```