assistant: export chat log to file (request, response, tool call input/output) (#8430)
### Summary

- addresses https://github.com/posit-dev/positron/issues/8033
- system prompts and session context will be addressed separately
https://github.com/posit-dev/positron/issues/8429
- leverages the handling for existing command `Chat: Export Chat...` to
include the tool call info and log directly to a new file in the
workspace; also enables file picker flow
- once the chat log is exported, a notification is shown with the log
file path and a button to open the file

### Release Notes

#### New Features

- Assistant: chat log can be exported for debugging/testing (#8033)

### QA Notes

@:assistant

#### QA Assistant testing notes

1. Set `"positron.assistant.toolDetails.enable": true` (includes tool
call details in the chat view, which allows the tool call details to be
logged in the exportable chat log as well)
2. Use Positron Assistant Chat
3. Run the command `positron-assistant.exportChatToFile` which will
create a JSON log file in the workspace with the assistant chat contents
4. Parse the request, response and tool call details from the JSON file
as needed

#### User export debug logs notes

1. Optionally set `"positron.assistant.toolDetails.enable": true`:
Enable tool call input and output details in the chat view. Must be
enabled to include tool call details in the chat export.
2. Chat in sidebar, inline in the editor or inline in the terminal
3. Run one the following commands to export the chat log for the last
focused chat widget
- `Positron Assistant: Export the current chat to a file in the
workspace`: exports the chat to a timestamped file in the workspace (not
available when in a workspace-less window)
- `Positron Assistant: Export the current chat to...`: choose the file
name and location to export the chat log to