Don't leak RUST_LOG into supervisor (#8541)
This change removes the `RUST_LOG` environment variable from the kernel
supervisor's environment. It is (now) unnecessary since we already use a
command-line argument to set the log level, and because kernels inherit
the supervisor's environment, it can affect the log level of Rust-based
programs invoked from the kernels (such as uv).

Addresses #8538.


### Release Notes

<!--
Optionally, replace `N/A` with text to be included in the next release
notes.
The `N/A` bullets are ignored. If you refer to one or more Positron
issues,
these issues are used to collect information about the feature or
bugfix, such
as the relevant language pack as determined by Github labels of type
`lang: `.
  The note will automatically be tagged with the language.

These notes are typically filled by the Positron team. If you are an
external
  contributor, you may ignore this section.
-->

#### New Features

- N/A

#### Bug Fixes

- Remove superfluous `RUST_LOG` environment variable from R and Python
sessions

### QA Notes

- It should still be possible to specify `RUST_LOG` manually via e.g.
the Environment Variables feature
- The R extension sets a value for `RUST_LOG` as well, so you'll see
that value in R sessions
- The main thing to verify here, in addition to `RUST_LOG` no longer
being overridden, is that the kernel's log level setting still works