Python: Safely and efficiently compute variable display values (#8545)
Addresses #8245. The approach is inspired by pydevd which backs the
debugger used by most Python IDEs today.

The focus is on safely (i.e. no hangs) and efficiently calculating and
sending display values over the wire, _not_ on perfectly formatting
those display values, hence the large max character values. The idea is
that further formatting would be handled in the UI, although we can
change this if needed.

I've mostly left the display formatting of the larger third party
objects as is (numpy, pandas, etc). The lower level changes here should
make refinements to those easier in future.

### Release Notes

#### New Features

- N/A

#### Bug Fixes

- Improved the performance of Python variable updates (#8245).

### QA Notes

See the repro steps in #8245 as well as the new unit tests.

@:variables @:data-explorer @:console

---------

Signed-off-by: Wasim Lorgat <mwlorgat@gmail.com>
Co-authored-by: Isabel Zimmerman <54685329+isabelizimm@users.noreply.github.com>